+++ src/main/java/com/mendonca/menssagerchat/service/UserMessengerService.java
@@ -1,5 +1,7 @@
 package com.mendonca.menssagerchat.service;
 
+
+import java.util.Map;
 import java.util.Optional;
 
 import com.mendonca.menssagerchat.model.UserMessenger;
@@ -7,5 +9,11 @@ import com.mendonca.menssagerchat.model.UserMessenger;
 public interface UserMessengerService {
 
 	Optional<UserMessenger> searchByUserName(String userName);
+	
+	public Map<String,Boolean> retreaveUsersAvalible(String userName);
+	
+	public void alterCurrentStatus(String userName, Boolean status );
+	
+	public boolean saveUser( UserMessenger userMessenger);
 	
 }